refact(cohort): load cohort by instance moved from instance manager to cohort lib
